requirements.txt
python-poppler
numpy
opencv-python 
matplotlib